Remove event count from day view
Looks like this was left in for debugging. It displays a random string
in the day view with no style hooks to hide it. On the example page, if
you scroll all the way up in the day view, the count shows right next to
"12am". Taking this out so that it doesn't interfere with the events
display.